requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
oauthlib==3.2.1
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.2.1
pytz==2022.2.1
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
